+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.6.0
+
+- support Redactor callbacks: uploadcareShow, uploadcareDone, uploadcareCancel
+- compatibility with latest Redactor 2 toolbar icons (default)
+- new setting: buttonIcon - use 'fa fa-picture-o' for backwards compatibility
+- restore editor selection on cancel
+- assign data-image and data-file attributes for Redactor storage management,
+  see [docs](https://imperavi.com/redactor/docs/storage/)
+
 ## 1.5.2
 
 - update CDN link to the widget